+++ src/formulafence/models.py
@@ -3134,6 +3134,7 @@ class FormulaCachedResultEntry:
     result_signature: str = field(repr=False)
     is_broken_reference_error: bool = False
     is_divide_by_zero_error: bool = False
+    is_numeric_error: bool = False
 
     @property
     def location(self) -> CellKey: